Runvm: more robust wait for VM startup.
Occasionally we saw that ssh into the guest would succeed, then immediately
after fail for a brief period before becoming stable and succeeding. This
caused build failures.

Work-around it by checking several times (with a small delay in-between)
that ssh into the guest is really up and stable before proceeding.