Buildbot: Use categories to distinguish between stable and unstable builders.
Go through all the builders in the Buildbot configuration, and place them
in one of just two categories: "main" or "experimental".

"main" is builders that must be kept green at all times. Any failure is critical.

"experimental" are for builders that people for one reason or another want to
keep in the config, but are not willing to ensure will not have any false
positives. Also useful for new builders until they are fully working.

Create a template for the grid.html display that has links to filter
different categories. If no category is specified, redirect to the "main"
category, to avoid presenting red builders by default.