Unify dir and ver variable names and var loading
Update scripts to all use the same variable names to refer to
directories and package versions.

Also unify how vars are loaded so that they all load from the same
location (~/.prep.conf) and don't rely on their own internally set
variables if they don't need to and changing a variable in one place
sets it for everyone.