Add Ubuntu 18.04 Bionic
This required several changes because this is the first Ubuntu release
to not support i386, so our default Ubuntu behavior of adding amd64 and
then i386 packages broke for bionic.

Solution was to split the i386 package section into it's own case
statement that only gets called for those ubuntu versions that actually
have i386 packages (trusty, xenial, and artful). This is similar to how
we handle arm64 and ppc64le packages.

While I was at it I updated those to use case statements as well. They
are much more readable than the if statements.