Fix authentication adjustment for systems with Debian-packaged MariaDB
Pre-10.4 versions of MariaDB packaged by Debian can have unix_socket
authentication by default, and they don't allow multiple authentication
methods yet. So, just setting a password is not enough for them, they
need to have the plugin field adjusted as well. In the fixed if,
the first option was earlier changed to run under sudo in order to
cover also pre-installed MariaDB 10.4 which doesn't need and wouldn't
allow changing the plugin field in mysql.user. But it broke the logic
for the above-mentioned pre-10.4 versions, because instead of failing
on "if" and falling through to "elif" with plugin adjustment, they
started passing the "if" which doesn't help them.

For now, the previous variant is restored, so that "if" works for
normal plugin-authenticated roots which just don't have the password
pre-configured, and the "elif" works for pre-10.4 Debian-packaged
MariaDB. If 10.4 is pre-installed, neither will work for it, but
so far this is just a hypothetical possibility, we don't have upgrade1
test for 10.4 as an "old" version. It will need to be covered later.