buildbot: ignore versioned library dependencies in srpm step
dependencies like "librt.so.1(GLIBC_2.2.5)(64bit)" are
reduced to "librt.so.1".

This could miss two kinds of differences. When the install host
has newer libraries than the build host - this is fine, and should
not fail the test. And when the build host has newer libraries than
the install host. This must be an error, if those libraries are
not available for install host. But then the install test will fail,
so this is covered.