buildbot: fix the galera check in deb install tests
it installs galera from http://yum.mariadb.org/galera/repo3/deb/ or
http://yum.mariadb.org/galera/repo4/deb/ - it must test whether such
a repo exists using a url it'll be installing from.

It used to test for http://yum.mariadb.org/galera/repo/deb/
and then replace "repo" with "repo3" or "repo4". Which was wrong
as the set of supported distros in "repo" and "repo4" is different.