Build BB containers in CI
This permits to build multi-arch containers in GH action CI. If check
passes, images are pushed to quay.io
(https://quay.io/repository/mariadb-foundation/bb-worker).

Two couples of secrets are needed for the CI to work:
- RHEL subscription credentials (only needed to build RHEL images);
- quay.io robot token.

If those are not provided, the CI will build and check non-rhel based
images but it will not try to push it to the quay.io registry.

TODO:
- rebuild only dockerfiles that have changed;
- schedule build.