bb.org: Add docker library cleanup on dependent failure
Container builds are based off 3 or 4 dependent builders triggering
the amd64-rhel8-docker builder to trigger the builds. These represent
the number of architectures that can be built depending on the major
version at the time.

The container and manifest are only pushed to quay.io once the full
complement of architectures is reached. The two contraints that
led to this approach are:
* you cannot push an image without a tag
* if you pushed a manifest of a particular tag, users fetching that
  tag of an with an architecture that hasn't finished building with
  receive a non-functioning image (for their architecture).

If a deb-autobake builder for ubuntu-2004 (or 1804 for 10.2 branch)
fails, this results in a number of images and their manifest waiting
for a trigger that won't happen.

As such we create a trigger step that happens alwaysRun=True, but only
runs if the current build step result is failure. This triggers the
cleanup script on the amd64-rhel8-dockerlibrary builder to remove
the image and manifests of those that won't complete.

"results" appears to be BuildStep attribute per
master/buildbot/process/buildstep.py in the buildbot code.

The unused(broken?) ifStagingSucceeding uses step.build.results in
('SUCCESS', 'WARNINGS'), so I'm unsure if which invokation is correct.