integrate processing new and old bb files together
Add functions to be able to process files from the old and new buildbots
in an automated way, including verifying that all required files are
present for each MariaDB release series. This includes the introduction
of a new -c <ci_tarbuildnum> variable to the script.

Also:

- change rsyncs to use --delete-before instead of --delete-after, which
helps avoid downstream sync clutter (syncing from primary mirrors to
secondary mirrors). Basically, get rid of the files we don't want sooner
in the process so that they don't hang around for possibly hours.

- Remove win32 builds. Those have been discontinued.