bb.org: fix galera timeout
Buildbot seems to enter systematically in timeout if it is not able to
pull logs from the libvirt worker. So, it's mandatory to fetch logs from
a readable and browsable (+x) PATH by the buildbot user (/var/log/syslog
is not working for instance).

Plus use of lazylogfiles that seems cleaner, see
https://docs.buildbot.net/current/manual/configuration/steps/shell_command.html.