update Docker Library Release
Because we don't want to be Naughty (from the official images test
(https://github.com/docker-library/official-images/runs/6542938100?check_suite_focus=true),

Our commits need to be in the master branch, so we merge origin/next
back to master now that the release is done.

Since the rests of these steps are in one hit, we can do all the
releases and test in one commit too.