Add start of GitDM based metrics
These metrics calculate contributor and contributor affiliation metrics
for commits. It currently uses a submodule to pull the Linux kernel
version of GitDM but this may well need to change as we customise it for
our usage.

There are some improvements to be made to this which are outlined in the
TODO.md. A contribution guide to the email mappings will eventually be
needed.