Prompt to copy previous releases on mirrors
Making a copy of the previous release on mirrors (corp and org) prior to
uploading a new release is a good way to speed up a given upload.
Instead of just blindly doing it for both the corp and org mirrors when
this option is chosen, ask to do it each time. This allows for better
flexibility when uploading.